+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## 0.0.14
-
- * new option `--permission`: Permission granted to the opened browser: e.g. clipboard-read, clipboard-write. See full list of options [here](https://playwright.dev/docs/1.14/emulation#permissions). Argument can be added multiple times.
+ * new option `--extensionPath` : A path pointing to a folder containing additional extensions to include. Argument can be provided multiple times.
+ * new option `--permission`: Permission granted to the opened browser: e.g. clipboard-read, clipboard-write. See full list of options [here](https://playwright.dev/docs/1.14/emulation#permissions). Argument can be provided multiple times.
  * new option `--hideServerLog`: If set, hides the server log. Defaults to true when an extensionTestsPath is provided, otherwise false.
  * close server when browser is closed